Move CScript class and dependencies to script/script
