+++ src/Makefile.am
@@ -98,6 +98,7 @@ BITCOIN_CORE_H = \
   rpcclient.h \
   rpcprotocol.h \
   rpcserver.h \
+  script/script.h \
   scriptutils.h \
   serialize.h \
   sync.h \
@@ -206,6 +207,7 @@ libbitcoin_common_a_SOURCES = \
   keystore.cpp \
   netbase.cpp \
   protocol.cpp \
+  script/script.cpp \
   scriptutils.cpp \
   $(BITCOIN_CORE_H)
