+++ src/qt/transactiondesc.cpp
@@ -11,7 +11,7 @@
 #include "db.h"
 #include "main.h"
 #include "paymentserver.h"
-#include "scriptutils.h"
+#include "script/script.h"
 #include "transactionrecord.h"
 #include "timedata.h"
 #include "ui_interface.h"
