+++ src/test/base58_tests.cpp
@@ -9,7 +9,7 @@
 #include "data/base58_keys_valid.json.h"
 
 #include "key.h"
-#include "scriptutils.h"
+#include "script/script.h"
 #include "uint256.h"
 #include "util.h"
