+++ src/test/multisig_tests.cpp
@@ -5,6 +5,7 @@
 #include "key.h"
 #include "keystore.h"
 #include "main.h"
+#include "script/script.h"
 #include "scriptutils.h"
 #include "uint256.h"
